requirements.txt
Pillow
ttkbootstrap